git subrepo pull (merge) kigs
subrepo:
  subdir:   "kigs"
  merged:   "3b60b377c2"
upstream:
  origin:   "https://github.com/assoria/kigs.git"
  branch:   "master"
  commit:   "a7444c5f48"
git-subrepo:
  version:  "0.4.1"
  origin:   "???"
  commit:   "???"

Former-commit-id: f5459927175a23a8aa72ef8ddbfb06164a6c0b90
Former-commit-id: 69541f016b85acf0a2e0cd9d16d66345c8320596

+++ framework/2DLayers/Sources/UI/UITextInput.cpp
@@ -45,7 +45,7 @@ void UITextInput::UpdateKeyboard(kstl::vector<KeyEvent>& keys)
 	if (!keys.empty() && HasFocus())
 	{
 		usString newString = mIsCueBanner ? usString("") : mText.const_ref();
-		newString.reserve(newString.length() + mKeycodeList.size());
+		newString.reserve((int)(newString.length() + mKeycodeList.size()));
 
 		for (auto& key : keys)
 		{
